requirements.txt
# LangChain core packages
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-postgres 
langsmith
langgraph
langchain_experimental

# Vector stores and embeddings
langchain-pinecone


# Document processing
pypdf

# Environment and utilities
python-dotenv

# OpenAI
openai

# FastAPI and web server
fastapi
uvicorn
pydantic

# Other dependencies
tqdm
psycopg2
psycopg2-binary
boto3
unstructured[pdf]
numpy
protobuf
pydantic-settings
google-search-results